+++ lib/clixon/clixon_xpath_ctx.h
@@ -76,6 +76,7 @@ struct xp_ctx{
     enum xp_objtype xc_type;
     cxobj         **xc_nodeset; /* if type XT_NODESET */
     int             xc_size;    /* Length of nodeset */
+    int             xc_position;
     int             xc_bool;    /* if xc_type XT_BOOL */
     double          xc_number;  /* if xc_type XT_NUMBER */
     char           *xc_string;  /* if xc_type XT_STRING */